apps/products/api_v1/serializers.py
from rest_framework import serializers

from apps.products.models import Product, Category


class CategorySerializer(serializers.ModelSerializer):

    class Meta:
        model = Category
        fields = ('id', 'name', 'description')


class ProductSerializer(serializers.ModelSerializer):

    class Meta:
        model = Product
        fields = ('id', 'category', 'name', 'active', 'price', 'short_description', 'long_description',)
